refactor: 책임 연쇄 패턴을 사용하도록 소스코드 리펙토링
