+++ src/test/java/calculator/handler/CalculatorHandler.java
@@ -1,4 +1,4 @@
-package calculator.chain;
+package calculator.handler;
 
 /**
  * <pre>
