+++ src/test/java/calculator/handler/CustomHandler.java
@@ -1,6 +1,5 @@
-package calculator.chain;
+package calculator.handler;
 
-import calculator.CalculratorValidation;
 import calculator.Numbers;
 import calculator.separator.CustomSeparator;
 import calculator.separator.Separator;
